Completed updates

+++ CLDV POE 1 Example Movie Database/Controllers/MoviesController.cs
@@ -63,6 +63,13 @@
             var movie = await _db.Movies.FindAsync(id);
             if (movie is not null)
             {
+                var hasScreenings = await _db.Screenings.AnyAsync(s => s.MovieId == id);
+                if (hasScreenings)
+                {
+                    ModelState.AddModelError(string.Empty, "Cannot delete this movie because it has scheduled screenings.");
+                    return View(movie);
+                }
+
                 _db.Movies.Remove(movie);
                 await _db.SaveChangesAsync();
             }
